Create the Visit model
This will be used to model a visit to a given page from a user (or an
anonymous user when `user_id` is nil). I have decided to create this
table with a distinct `visited_at` column as the default rails
timestamps may be unreliable if database backfills are performed in the
future.