Add POST /visits endpoint
This commit adds the Visits controller and its related request spec. In
order to test the new controller action, I've had to add the POST
endpoint to the routes file and configure rspec to use route helpers.

The controller will accept nested parameters for the associated IpInfo
record. To account for IP Addresses that don't already exist on my
database, I've added a `before_validation` callback to the model that
will use the `.find_or_create_by` method so that a new IpInfo record is
created when a POST request is made using a new IP address.