Add support for requests from ngrok URLs
It isn't possible to collecting users' public IP addresses from
localhost so it is necessary to set up a ngrok tunnel when running the
frontend application in the development environment. As such, I will
need to permit API requests from both the ngrok free app domain and
localhost in the environment configuration.