Get IpInfo location during after_create
After new records are created, I would like to make a request to the
ipinfo.io API in order to persist geolocation data for the IP Address.
In order to do this, I have added an after_create hook to the IpInfo
model that will call the IpLocation service. This has meant refactoring
a number of tests in order to stub out the new API request made wherever the
`ip_info` factories are used.

I've also added a `#loopback?` method that will return true if non-public
IP address is stored and refactored the IpLocation service so that only
one API request is needed to obtain the coordinates and the VPN status.