Add tests for IpInfo after_create
This commit adds a set of tests for the after_create hook that is run
after IpInfo records are created. This was slightly challenging due to
the mocking that was needed for instances of the IpLocation service. I
am a little concerned that the resulting tests might prove brittle in
the long-run.