Add /pages index and serialiser
This commit adds the Pages controller and its `#index` action, the
response for which will be used in the dropdown list on the frontend.
This also adds the PageSerializer which is used to format the page
objects in the response body. The way in which rails handles
serialisation seems to have changed significantly over the years. I have
decided to create a serialiser that exists outside of the Page model
because I believe that this maintains a separation of concerns.