Support query parameters in the visits index
This commit adds support for query parmeters that can filter the list of
visits returned by the page ID, a start time (`from`) and an end time
(`to`). When they're missing, the "from" time will default to year 0
whilst the "to" time will default to the current time.

I've also added a new `.for_page` scope to the Visit model which is
used to apply the `page_id` filter in the `#index` controller action.