Simplify the Visit factory
I've noticed that a lot of tests fail because of the associations that
this factory builds by default. The main offender seems to be the associated
IpInfo record which has a uniqeness constraints on the address that
causes issues when duplicates are created. Moving the `ip_info`
association to a trait seems like the best solution for now. As the
`user` isn't required, I think I can remove this association completely.