Refactor the Visit specs
This commit adds tests for the before_validation callback that
associates a Visit with a new or existing IpInfo record. Originally, I'd
thought it easier to test this at the controller level in the request
spec but as the callback is triggered in the model, the visits specs
seem incomplete without them.

I've also reorganised the file more generally, placing scopes (class
methods) at the top of the file and tests related to instance methods
closer to the bottom.