Accept nested attributes for `User#registration_ip_info`
In order to store the user's first IP address, `User.create` must accept
the IP address as a nested attribute. This commit adds this alongside a
`#find_or_create_ip_info` before_validation (similar to the Visit
model).
I have also refactored some of the related unit tests as some of the
older tests weren't using a factory already which caused issues where
the IpInfo association was missing or duplicate.