Override sign_in and add /users/auth endpoint
By default, the `#sign_in` action actually throws an error when users
fail authentication which I don't think is very helpful. I've instead
overridden it so that a JSON response with an error is returned when
this happens.
As I don't want to rely on JWTs in this application (for security
reasons), I will need an extra endpoint to authenticate users by
checking their session cookies. Luckily, Devise provides helper methods
for this which I have used in my new endpoint.