Migration for adding an `is_admin` column to `users`
I want to be able to grant admin users access to certain pages in the
frontend and so will need to store this as a boolean. I believe that
actually granting users admin permissions in the UI is out-of-scope for
this task so this column will need to be set manually in the database
for any users who require it.