Add `/users/:user_id/visits` endpoint
I've decided to add an extra endpoint for the user visits instead of
adding this to the existing `/user/:id` endpoint because I expect that I
may want to paginate visits in the future and that would be simpler to
do on a dedicated controller action.
This endpoint repurposes the `#index` action for the existing Visits
controller so I've had to refactor the controller slightly to take this
into account. Most of the filter logic for the action has been moved
into private methods to improve readability. For the changes in the
controller to work (particularly "User.find(params[:user_id]).visits"),
it was necessary to add a `has_many` relationship to the User model.