Minor refactors following rubocop
Attempts to refactor the `#filter_by_time` private method in the
VisitsController haven't fixed the branch condition warning that rubocop
has raised so I've decided to add the controller as an example. This
commit also renames the `#get_location_data` method which actually wrote
to the database, making the old name a little misleading. It will
now be called `#set_location_data`.